selftests/landlock: Add realworld workload based on find tool
Implement script that measures Landlock overhead for workload in which
find tool is executed on Linux source code folder. This workload is tested
with 5, 10 depth values and few number of layers.

This workload is useful to measure Landlock overhead under different
number of layers and different keys of the filesystem ruleset.

Signed-off-by: Mikhail Ivanov <ivanov.mikhail1@huawei-partners.com>