User-space API for creating a supervisor-fd
We allow the user to pass in an additional flag to
landlock_create_ruleset which will make the ruleset operate in
"supervise" mode, with a supervisor attached. We create
additional space in the landlock_ruleset_attr structure to pass
the newly created supervisor fd back to user-space.

The intention, while not implemented yet, is that the user-space
will read events from this fd and write responses back to it.

Note: need to investigate if fd clone on fork() is handled
correctly, but should be fine if it shares the struct file. We
might also want to let the user customize the flags on this fd,
so that they can request no O_CLOEXEC.