Reference supervisor from layers
We store a nullable reference to the supervisor in each
landlock_layer of each landlock_rule. Later we will implement
generating and sending events to the correct supervisors.

The refcount situation here is a bit tricky, we stick to always
incrementing it when we pass a reference from one function to
another.