Enhance the sandboxer example to support landlock-supervise
This is perhaps a bit overengineered with the ppoll on 5 different fds,
but it makes sure the sandboxed child can't try to print anything to the
terminal from a different thread while an access request is pending
(otherwise it could trick the user by printing over the request text).
This also makes sure inputs are directed to the right place (the child
when no prompt, or the sandboxer itself when an access request is shown).

But even with that, I'm not claiming this "sandbox" with supervise mode is
in any way production quality.  It's intended as a PoC.

Signed-off-by: Tingmao Wang <m@maowtm.org>