Define the supervisor and event structure
We define a refcounted structure that represents a supervisor,
and will later be exposed to the user-space via a file
descriptor.  Each supervisor has an event queue, where each event
is a access request by a landlock process that the supervisor
needs to make a decision on. The details of the event structure
will be populated in a future patch.

The struct is called landlock_supervise_event_kernel so that the
uapi header can use the shorter name.