[stash] todo
the event queue needs to be an xarray instead of just a list
also don't store landlock_supervisor reference in landlock_layers and end up duplicating the reference for each rule. A supervisor belong to layers, not rule, and it's not clear which "rule" triggers a supervisor event.
Imagine a layer denying everything and thus no rules...