fs/9p: add default option for path-based inodes
We're now re-using inodes in uncached mode as well.  Add an extra option
to control the inode identification behaviour, and make path-based the
default (so that things doesn't break even when the server sends the same
qid for different files).

In a future version, if we can negotiate with the server and be sure that
it won't give us duplicate qid.path, the default for those cases can be
qid-based.

Signed-off-by: Tingmao Wang <m@maowtm.org>