fs/9p: add default option for path-based inodes
Add an extra option to control the inode identification behaviour, and
make path-based default.  The reasoning for this default is that the
current 9pfs already does not re-use inodes even for same qid.path, but
because of the dcache, a file kept open by a process (by e.g. open, or
having the directory as CWD) means that other processes will re-use that
inode.  If we identify inode by path by default, this behaviour is
consistent regardless of whether files are kept open or not.

In a future version, if we can negotiate with the server and be sure that
it won't give us duplicate qid.path, the default for those cases can be
qid-based (possibly re-merging commit 724a08450f74 ("fs/9p: simplify iget
to remove unnecessary paths")...?)

Signed-off-by: Tingmao Wang <m@maowtm.org>