Add hlist_node member to struct landlock_rule
This is to prepare for the new domain. Since a rule can only be in either
a ruleset (rbtree based) or a landlock_domain (hashtable based), we can
make the node/hlist member of a union.  For now let create_rule initialize
is as before.

(Alternatively, if we use hashtable for both cases, then we save 8 bytes
per each rule, but then we will need some kind of resizable hashtable)

(Maybe we should use the relativistic hash tables after all?)

Signed-off-by: Tingmao Wang <m@maowtm.org>