landlock/domain: Define alloc and free
We will eventually need a deferred free just like we currently have for
the ruleset, so we define it here as well.  To minimize the size of the
domain struct before the rules array, we separately allocate the
work_struct (which is currently 72 bytes) and just keep a pointer in the
domain.

Signed-off-by: Tingmao Wang <m@maowtm.org>