landlock/domain: Define structure and macros for flat-array domains
This implements the structure proposed in in [1], using a flat array to
store the rules and eventually using binary search for queries.  The array
is stored in the domain struct itself to avoid extra pointer indirection
and make all the rule data as cache-local as possible.  The non-array part
of the domain struct is also kept reasonably small.  This works well for a
small (10 or 20 rules) ruleset, which is the common case for Landlock
users, and still has reasonable performance for large ones.

This will eventually make landlock_rule/landlock_ruleset only needed for
unmerged rulesets, and thus it doesn't have to store multiple layers etc.
create_rule and insert_rule would also hopefully become much simpler.

Link: https://lore.kernel.org/all/20250526.quec3Dohsheu@digikod.net/ [1]

Signed-off-by: Tingmao Wang <m@maowtm.org>