landlock/domain: Implement merging walk of a parent domain and a ruleset
This implements a "merging" walk function, which will make merging domains
in a later commit very straightforward.  While the function has a more
than average number of arguments, the goal is that the caller will be able
to easily use it for both fs and net, and also use the same function both
to pre-calculate the number of indices and layers needed, and then to
actually write them out.

Signed-off-by: Tingmao Wang <m@maowtm.org>