landlock: Set the max rules limit in a domain to U16_MAX.
We need to reduce the limit from U32_MAX as we use u32 for various
landlock_domain_index or landlock_layer indices.

On Mon, 2 Jun 2025 at 21:50:05 +0200, Mickaël Salaün wrote [1]:
> Correct.  We can either use u64 or reduce the maximum number of rules.
> I think LANDLOCK_MAX_NUM_RULES set to U16_MAX would be much more than
> the worse practical case.  Even if one buggy policy tries to add one
> rule per network port, that will be OK.  We could even reasonably test
> this limit.  We'll need to backport this change but I'm OK with that.

Note that a limit of 2^24 still leaves us with more than enough room even
for u32 indices, but for future-proofing and given agreement with Mickaël,
setting this to U16_MAX here.

Link: https://lore.kernel.org/all/20250602.uBai6ge5maiw@digikod.net/ [1]

Signed-off-by: Tingmao Wang <m@maowtm.org>