selftests/landlock: Support running the full fs test suite on another FS
Adds a TMP_BIND_TO environment variable which the test binary will check,
and if present, instead of mounting a tmpfs on ./tmp, it will bind mount
that path to ./tmp instead.

Currently there is the layout3_fs tests which runs a few tests (but not
the full set of Landlock tests) in separate filesystems, notably no file
creation/write/rename etc.  This is necessary for certain special fs such
as proc or sysfs, as the tests can only read a specific path.  However,
for a more typical fs like v9fs, this is limitting.

This test makes it possible to run the full set on any filesystem (even
though this is still not automated).  Note that there are some expected
failures, such as v9fs not supporting RENAME_EXCHANGE, as well as the
known issue of ephemeral inodes, which may be fixed by a later revision of
[1].

Link: https://lore.kernel.org/v9fs/cover.1743971855.git.m@maowtm.org/ [1]
Signed-off-by: Tingmao Wang <m@maowtm.org>