selftests/landlock: Add rename tests for layout3_fs
Since the path-based inode identification feature needs special handling
for renames, we add some Landlock tests for renames on custom filesystems.

As part of this, also standardize the file path used in existing
layout3_fs variants to be file1_s1d1 (unless they are per-fs fixed paths).
This is so that we can rename this file to e.g. file2_s1d1 or file1_s2d1
(by renaming the parent dir).  This is an optional change - we could
always add an extra file creation in fixture setup for layout3_fs, but
since there's no particular reason to use "dir/file" as the test file, we
just make it use the "standard" name used in the rest of this fs_test.

To: "Mickaël Salaün" <mic@digikod.net>
To: "Günther Noack" <gnoack@google.com>
Signed-off-by: Tingmao Wang <m@maowtm.org>