landlock: Add a place for flags to layer rules
To avoid unnecessarily increasing the size of struct landlock_layer, we
make the layer level a u8 and use the space to store the flags struct.

Signed-off-by: Tingmao Wang <m@maowtm.org>
---

Note: conflict with 9a868cdbe6 ("landlock: Fix handling of disconnected
directories") on mic/next but is not hard to fix - need to backup the
collected rule flags too.

Changes since v1:
- Comment changes